# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A metrics collection and alerting server for storing and retrieving counter and gauge metrics",
        "title": "Metrics Collection API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "Retrieve an HTML page displaying all stored counter and gauge metrics",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "metrics"
                ],
                "summary": "Get all metrics summary",
                "responses": {
                    "200": {
                        "description": "HTML table with all metrics",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Template failure",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "Check service health and database connectivity",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "pong",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/update/": {
            "post": {
                "description": "Store a single counter or gauge metric with validation and optional persistence",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "metrics"
                ],
                "summary": "Store a single metric",
                "parameters": [
                    {
                        "description": "Metric data (counter with delta or gauge with value)",
                        "name": "metric",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Metrics"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request - Invalid JSON or unknown metric type",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - Empty metric ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Storage failure",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/update/counter/{metric}/{value}": {
            "post": {
                "description": "Legacy endpoint to store a counter metric via URL path parameters",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "deprecated"
                ],
                "summary": "Store counter metric (deprecated)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Metric name",
                        "name": "metric",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Counter value (integer)",
                        "name": "value",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request - Invalid value format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - Empty metric name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Storage failure",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/update/gauge/{metric}/{value}": {
            "post": {
                "description": "Legacy endpoint to store a gauge metric via URL path parameters",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "deprecated"
                ],
                "summary": "Store gauge metric (deprecated)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Metric name",
                        "name": "metric",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Gauge value (float)",
                        "name": "value",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request - Invalid value format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - Empty metric name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Storage failure",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/updates/": {
            "post": {
                "description": "Store an array of metrics (counters and gauges) in a single batch operation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "metrics"
                ],
                "summary": "Store multiple metrics in batch",
                "parameters": [
                    {
                        "description": "Array of metrics to store",
                        "name": "metrics",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.Metrics"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request - Invalid JSON format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - Empty metric ID or empty array",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Storage failure",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/value/": {
            "post": {
                "description": "Retrieve the current value of a counter or gauge metric",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metrics"
                ],
                "summary": "Get a metric value",
                "parameters": [
                    {
                        "description": "Metric identifier (id and type required)",
                        "name": "metric",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Metrics"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metric with current value (delta for counter, value for gauge)",
                        "schema": {
                            "$ref": "#/definitions/dto.Metrics"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid JSON or unknown metric type",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - Empty metric ID or metric not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Retrieval failure",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/value/counter/{metric}": {
            "get": {
                "description": "Legacy endpoint to retrieve a counter metric value via URL path",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "deprecated"
                ],
                "summary": "Get counter metric (deprecated)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Metric name",
                        "name": "metric",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Counter value as plain text",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - Empty metric name or metric not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Retrieval failure",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/value/gauge/{metric}": {
            "get": {
                "description": "Legacy endpoint to retrieve a gauge metric value via URL path",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "deprecated"
                ],
                "summary": "Get gauge metric (deprecated)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Metric name",
                        "name": "metric",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gauge value as plain text",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - Empty metric name or metric not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Retrieval failure",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.Metrics": {
            "type": "object",
            "properties": {
                "delta": {
                    "description": "Delta holds the counter value increment. Non-nil only for counter metrics.",
                    "type": "integer"
                },
                "id": {
                    "description": "ID is the unique name/identifier of the metric.",
                    "type": "string"
                },
                "type": {
                    "description": "MType specifies the metric type: \"counter\" or \"gauge\".",
                    "type": "string"
                },
                "value": {
                    "description": "Value holds the gauge value. Non-nil only for gauge metrics.",
                    "type": "number"
                }
            }
        }
    }
}